Add files via upload
The css file and the index are only the ones you need if any changes needed. The rest is a library, don't change anything there.

The library is used specific to design and generate a html page to pdf format. I made sure that the html page won't download directly to the  computer as a pdf format, instead it will change to a pdf page and then within that page the user will be able to download to their computer. This will make the user to see how the pdf format will look like before downloading it to the computer and even if the user wants to print it instead of downloading it.